requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apispec==6.8.1
asttokens==3.0.0
audioop-lts==0.2.1
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-speech==2.31.1
google-cloud-texttospeech==2.25.1
google-genai==1.5.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.0b2
pydantic_core==2.32.0
pydub==0.25.1
Pygments==2.19.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==76.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.39
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webargs==8.6.0
websockets==14.2
Werkzeug==3.1.3
